251/objhits: de-parameterize a scan helper; narrow two angle deltas
PressureSwitchFB_update 99.65087 -> 99.76309: scanTrackedSlots carried
three in-parameters (slotIndex, foundTrackedObject, emptyValue) that every
call site passed 0 -- reconstruction artifacts. Reducing the signature to
(stateAddress) with those as proper locals fixes the scratch band inside the
slot-scan loop, since the inlined helper's params were modelled as a
distinct register population from plain locals. Stacked with a load-class
decl-order fix (positionAddress ahead of trackedObject).

objJointTracksAimAtTarget 99.65466 -> 99.81982: d1/d2 are s16-valued joint
angle differences, and retail's extsh is the declared width of the local,
not a cast artifact. Widening them to int inflated their webs across the
yaw/pitch limiting arithmetic. A 16-way type sweep shows d1 and d2 are
jointly load-bearing while the neighbouring temps are inert.

Units 99.84553 / 99.80040. Probe-verified; every sibling bit-identical and
data sections unchanged in both units.

+++ src/main/objhits.c
@@ -4942,9 +4942,9 @@ s16 objJointTracksAimAtTarget(GameObject* obj, GameObject* target, f32* pos, u8*
             p4 += 0x60;
         } else {
             s16* fv = found[0];
-            int d1 = (s16)((s16)((fv[1] + dst[0]) >> 1) - fv[1]);
+            s16 d1 = (s16)((s16)((fv[1] + dst[0]) >> 1) - fv[1]);
             s16 lim;
-            int d2;
+            s16 d2;
             int t2;
             int lim3;
